WIP change to more flexible context.
Rather than constrain the terminal nodes to lists and sets for the
data and accumulator, use a single generic type to provide a "context"
to the terminal nodes. Then the nodes can do as they see fit with the
context and adds a lot more flexibility to what you can do inside and
AST node.